added JwtTokenService and new IdentityService methods for architectural purity

+++ src/Application/ApplicationUsers/Commands/DeleteMyAccount.cs
@@ -9,24 +9,25 @@ public record DeleteMyAccountCommand : IRequest;
 
 public class DeleteMyAccountCommandHandler : IRequestHandler<DeleteMyAccountCommand>
 {
-    private readonly IApplicationDbContext _context;
     private readonly IIdentityService _identityService;
     private readonly IUser _user;
     private readonly IStripeCustomerService _stripeCustomerService;
 
-    public DeleteMyAccountCommandHandler(IApplicationDbContext context, IIdentityService identityService, IUser user, IStripeCustomerService stripeCustomerService)
+    public DeleteMyAccountCommandHandler(IIdentityService identityService, IUser user, IStripeCustomerService stripeCustomerService)
     {
-        _context = context;
         _identityService = identityService;
         _user = user;
         _stripeCustomerService = stripeCustomerService;
     }
 
-    public async Task Handle(DeleteMyAccountCommand request, CancellationToken cancellationToken)
+    public async Task Handle(DeleteMyAccountCommand command, CancellationToken cancellationToken)
     {
-        var user = await _context.ApplicationUsers
-            .AsNoTracking()
-            .FirstOrDefaultAsync(au => au.Id == _user.Id);
+        var user = await _identityService.FindUserByIdAsync(_user.Id!);
+
+        if (user == null)
+        {
+            throw new UnauthorizedAccessException();
+        }
 
         var result = await _identityService.DeleteUserAsync(user!);
